Create file usage utility command and service
